refactor: Move streamer mode handling to separate model
