+++ common/src/main/java/com/wynntils/functions/WorldFunctions.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © Wynntils 2022-2025.
+ * Copyright © Wynntils 2022-2026.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
 package com.wynntils.functions;
@@ -105,7 +105,7 @@ public class WorldFunctions {
     public static class InStreamFunction extends Function<Boolean> {
         @Override
         public Boolean getValue(FunctionArguments arguments) {
-            return Models.WorldState.isInStream();
+            return Models.StreamerMode.isInStream();
         }
 
         @Override
